moves cursor to correct position on empty lines
